Add files via upload
The game narrates the story of a soccer player who is having his last chance to show himself and get promoted, the user will play the role of this player and have to make a decisions to complete the goal, based on decisions user chosen the end of the story will vary. The game has some open ended questions along with questions with options. Good Luck!